MA-8204 RFC: disable pcm_close in standby to reduce output latency in sabresd
This is a workaround fix for the output latency in sabresd board,
which uses the wm8962 audio codec.
cold output latency for wm8962 is high for there is a delay in
wm8962_set_bias_level, see:
http://wiki.freescale.net/display/MPUSW/wm8962+play+latency

In output_standby, if matches wm8962, disable pcm_close to remove
added latency by the next pcm_open.
Need to consider whether there need a force standby, e.g. switch from
wm8962 to hdmi in the same primary stream.

Change-Id: Ibc0f3a2123f430a03673e8d55f779d8d999fc347
Signed-off-by: Winter Wang <wente.wang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/196
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: fang hui <hui.fang@nxp.com>